// src/app/pages/hotel/layout/layout-list/shared/layout-list.enum.ts
export enum LayoutListEnum {

    totalLayoutText = 'Total Number of Rooms: ',

    roomHeaderName = 'Room Name',
    numberOfThisTypeHeaderName = 'Number of this Type',
    sizeHeaderName = 'Room Size',
    measureUnitHeaderName = 'Measure',
    priceHeaderName = 'Price',
    operationsHeaderName = 'Operations',

    dataSourceVariableName = 'dataSource',


    priceColumn = 'price',
    roomNameColumn = 'roomName',
    quantityColumn = 'quantity',
    totalRoomMeasureColumn = 'totalRoomMeasure',
    unitMeasureRoomColumn = 'unitOfMeasurementName',
    operationsColumn = 'operation',

    idJsonNode = 'id',
    roomJsonNode = 'roomInfo',
    roomNameJsonNode = 'roomName',
    nameJsonNode = 'name',
    customNameJsonNode = 'customName',

    quantityJsonNode = 'quantity',

    
    deleteConfirmMessage = 'Are you sure you want to delete the property "&a"?',
    tagToReplace = '&a',
    deleteLayoutErrorMessage = 'There was an error trying to delete your layout information',
}